requirements.txt
pysaml2==2.4.0
python==2.7
defusedxml==0.4.1
lxml
